+++ xxl-job-core/src/main/java/com/xxl/job/core/biz/ExecutorBiz.java
@@ -37,7 +37,7 @@ public interface ExecutorBiz {
      * @param fromLineNum
      * @return
      */
-    public ReturnT<LogResult> log(long logDateTim, int logId, int fromLineNum);
+    public ReturnT<LogResult> log(long logDateTim, long logId, int fromLineNum);
 
     /**
      * run
